requirements.txt
fastapi
uvicorn[standard]
websockets
trafilatura
curl_cffi
python-dotenv
